Do not rename test files on-the-fly to share functions
Instead this commit moves some functions that could be used by fuzzers.
